requirements.txt
cloudshell-shell-core>=3.1.0,<3.2.0